exercises 4.15-4.22

+++ part4/bloglist/controllers/blogs.js
@@ -1,27 +1,47 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const jwt = require('jsonwebtoken')
+const mw = require('../utils/middleware')
+
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})  
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
-  })
-  
-blogsRouter.post('/', async (request, response) => {
-    const body = request.body
-    const blog = new Blog({title: body.title, author: body.author, url: body.url, likes: body.likes || 0})
-    const savedBlog = await blog.save()
-    response.status(201).json(savedBlog)
-  })
+})
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.post('/', mw.userExtractor, async (request, response) => {
+  const body = request.body
+  console.log(request.user)
+  const user = request.user
+  if (!user) {
+    return response.status(401).json({ error: 'Not authorized' })
+  }
+  const blog = new Blog({ title: body.title, author: body.author, url: body.url, likes: body.likes || 0, user: user._id })
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  response.status(201).json(savedBlog)
+})
+
+blogsRouter.delete('/:id', mw.userExtractor, async (request, response) => {
+  console.log(request.headers.authorization)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'invalid token' })
+  }
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== decodedToken.id) {
+    return response.status(401).json({ error: 'Not authorized to delete the blog' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
   const body = request.body
-  const blog = {title: body.title, author: body.author, url: body.url, likes: body.likes}
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true, runValidators: true })
+  const blog = { title: body.title, author: body.author, url: body.url, likes: body.likes }
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true, runValidators: true, context: 'query' })
   response.json(updatedBlog)
 })
